refactor: 회원가입 및 로그인 js 파일에서 중복되는 함수 및 변수 member.js로 통합
